Separate Inference and Management Control Ports
