+++ benchmarks/config_template.properties
@@ -1,3 +1,5 @@
 model_server_home=/mxnet-model-server/
 model_store=/mxnet-model-server/frontend/modelarchive/src/test/resources/models/
 load_models=noop-v0.1,noop-v1.0
+address=0.0.0.0:8080
+management_address=0.0.0.0:8081
