+++ frontend/server/src/test/resources/config.properties
@@ -1,5 +1,6 @@
 # debug=true
-# port=8080
+inference_address=https://127.0.0.1:8443
+management_address=https://127.0.0.1:8444
 model_server_home=../..
 model_store=../modelarchive/src/test/resources/models
 load_models=noop-v0.1,noop-v1.0
@@ -8,10 +9,8 @@ load_models=noop-v0.1,noop-v1.0
 # job_queue_size=1000
 # number_of_gpu=1
 # netty_io_ratio=50
-use_ssl=true
 # keystore=src/test/resources/keystore.p12
 # keystore_pass=changeit
 # keystore_type=PKCS12
 private_key_file=src/test/resources/key.pem
 certificate_file=src/test/resources/certs.pem
-
